Rename Item class to Task
